Fix bugs

+++ Workouts/Models/Set+CoreDataProperties.swift
@@ -17,7 +17,7 @@ extension Set {
     }
 
     @NSManaged public var reps: Int16
-    @NSManaged public var weight: Int16
+    @NSManaged public var weight: Float
     @NSManaged public var exercise: Exercise?
 
 }
